backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.0.0
